library/core/tests/iter rearrange & add back missed doc comments
